chore: move assets list out into its' own file
