+++ frontend/routes/auth/wallets/key-details/assets-list/assets-list.tsx
@@ -1,62 +1,14 @@
-import { vegaAccountType } from '@vegaprotocol/rest-clients/dist/trading-data'
 import { AsyncRenderer } from '../../../../../components/async-renderer/async-renderer'
 import { SubHeader } from '../../../../../components/sub-header'
 import { VegaSection } from '../../../../../components/vega-section'
-import { isActiveMarket } from '../../../../../lib/markets'
-import { useAssetsStore } from '../../../../../stores/assets-store'
-import { useMarketsStore } from '../../../../../stores/markets-store'
 import { useAccountsStore } from './accounts-store'
 import { AssetCard } from './asset-card'
 import { useAccounts } from './use-accounts'
 import { Notification, Intent } from '@vegaprotocol/ui-toolkit'
+import { AssetListEmptyState } from './asset-list-empty-state'
 
 export const locators = {
   assetListDescription: 'asset-list-description'
-}
-
-const AssetListEmptyState = ({ publicKey }: { publicKey: string }) => {
-  const { assets } = useAssetsStore((state) => ({
-    assets: state.assets
-  }))
-  const { getMarketsByAssetId } = useMarketsStore((state) => ({
-    getMarketsByAssetId: state.getMarketsByAssetId
-  }))
-  const sortedAssets = assets
-    .map((asset) => ({
-      asset,
-      markets: asset?.id ? getMarketsByAssetId(asset.id).filter(isActiveMarket) : []
-    }))
-    .sort((a, b) => {
-      return a.markets.length - b.markets.length
-    })
-
-  const top2Assets = sortedAssets.slice(0, 2)
-
-  return (
-    <div>
-      <SubHeader content="Balances" />
-
-      <p className="text-vega-dark-400 my-3">Currently you have no assets.</p>
-      {top2Assets.map(({ asset }) => {
-        if (!asset.id) return null
-        return (
-          <AssetCard
-            allowZeroAccounts={true}
-            key={asset.id}
-            accounts={[
-              {
-                balance: '0',
-                asset: asset.id,
-                owner: publicKey,
-                type: vegaAccountType.ACCOUNT_TYPE_GENERAL
-              }
-            ]}
-            assetId={asset.id}
-          />
-        )
-      })}
-    </div>
-  )
 }
 
 export const AssetsList = ({ publicKey }: { publicKey: string }) => {
